How do I download my Pull request? (#1)
* 0.14.1 Update For client and server

In this pull request I will be updating the PocketMine Server MINECRAFT_VERSION (Which will be the MC version the server accepts) & MINECRAFT_VERSION_NETWORK (Which will be the MC version the server runs on). It will make MC version 0.14.1 compatible with this server and, MC 0.14.1.

* Update PocketMine.php

* Added $logger->notice code

Added $logger->notice("Be sure to use MCPE v0.14.1 to join the server"); after const MINECRAFT_VERSION loads and added $logger->("Also, this is the MCPE version that the server uses. :)"); after MINECRAFT_VERSION_NETWORK loads in startup code.

* Fixes a Critical Bug/Also still adds $logger strings

Fixes the version bug + Also adds $logger strings like $logger->notice("Also, this is the MCPE version that the server uses. :)"); and, $logger->notice("Be sure to use MCPE v0.14.1 to join the server.");

* Changed Network Protocol 45 to 46

Changed Network Protocol 45 to 46.

* String bug fix which crashes the server entirely

String bug fix which crashes server entirely when it sees I am making a $logger string of text it gets confused because it's not logging the server on it's setting up variables and "consts" instead.

* Gets rid of unnecassary strings of text

Gets rid of unnecassary strings of text